将 ReplaceTextPatches 分成多个独立的文件，并将其中的功能函数独立到文件中
