+++ Localyssation.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35327.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Localyssation", "Localyssation.csproj", "{356033B9-C816-43E0-9C52-5CCD6611F8ED}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{356033B9-C816-43E0-9C52-5CCD6611F8ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{356033B9-C816-43E0-9C52-5CCD6611F8ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{356033B9-C816-43E0-9C52-5CCD6611F8ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{356033B9-C816-43E0-9C52-5CCD6611F8ED}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {90E1C35B-2A0A-46B0-8919-D37627620CE2}
+	EndGlobalSection
+EndGlobal
