+++ Patches/GameLoadPatches.cs
@@ -1,4 +1,5 @@
 using HarmonyLib;
+using Localyssation.Patches.ReplaceText;
 using System;
 using System.Collections;
 using System.Collections.Generic;
